Add files via upload
this is only the ROM

it can be played in gameboy and emulators


		BLOCK LEGEND

	You are a turist in some ruins in 
central america. You get lost and fall inside 
one of the piramid ruins. on the inside there 
are monsters and weird blocks. Push the blocks
to get to the doorway. This doorway will take
you to the next level. Complete all levels and 
you'll have escaped from the piramid you are 
trapped on.


----------------------------------------------

controls:

arrows to move
A to push blocks
B to use the special power
SELECT to reset level in case you mess up
START to pause level


----------------------------------------------

blocks:

[ ] common block: You can push it arround. There
	are 3 variations.
[?] ? block: These blocks will disappear after 
	pushing them. If something is blocking 
	the way, then it wont move or disappear
[G] B power block: If you push this block, it
	will disappear leaving a weird tile on 
	the floor. You can now press B to make 
	a wall or block disappear.
[!] ! block: these blocks explode clearing a whole
	area of 3x3
[%] weird sign block: push it around like common 
	blocks, but if it gets pushed over a sign 
	in the floor, it will turn into a doorway
[9] number blocks:Each time you push them, they 
	become a smaller number. if it reaches 0,
	you wont be able to move it again.

notes:
	Some walls look like one of the variations
of the common block.

	Some walls are invisible. These arent very 
common. Can you find one?

	Sometimes blocks can behave in strange ways...

------------------------------------------------------
Enemies:

	there are 3 bosses and one extra enemy
can you beat them all?

------------------------------------------------------

´´´Copyright (c) 2021 Jonatán Hidalgo Morales <jona32u4hm@gmail.com>

You are free to download and use this game and source code for free 
but you do not have the rights to re-distribute or sell this game in 
any way without explicit permission. ```